+++ packages/stack/src/views/Stack/CardStack.tsx
@@ -142,7 +142,7 @@ const getHeaderHeights = (
     const style = StyleSheet.flatten(headerStyle || {});
 
     const height =
-      typeof style.height === 'number'
+      'height' in style && typeof style.height === 'number'
         ? style.height
         : previous[curr.route.key];
 
@@ -614,16 +614,23 @@ export default class CardStack extends React.Component<Props, State> {
             const headerHeight =
               headerShown !== false ? headerHeights[route.key] : 0;
 
-            const { backgroundColor: headerBackgroundColor } =
-              StyleSheet.flatten(headerStyle) || {};
-
             let headerDarkContent: boolean | undefined;
 
             if (headerShown) {
               if (typeof headerTintColor === 'string') {
                 headerDarkContent = Color(headerTintColor).isDark();
-              } else if (typeof headerBackgroundColor === 'string') {
-                headerDarkContent = !Color(headerBackgroundColor).isDark();
+              } else {
+                const flattenedHeaderStyle = StyleSheet.flatten(headerStyle);
+
+                if (
+                  flattenedHeaderStyle &&
+                  'backgroundColor' in flattenedHeaderStyle &&
+                  typeof flattenedHeaderStyle.backgroundColor === 'string'
+                ) {
+                  headerDarkContent = !Color(
+                    flattenedHeaderStyle.backgroundColor
+                  ).isDark();
+                }
               }
             }